+++ CHANGELOG.md
@@ -4,7 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [V00g] - 2024-11-05
+
+## [V00h] - 2024-11-07
+
+### Added
+
+- improved movegen performance
+- improved performance in quiescence search
+
+### Fixed
+
+### Changed
+
+### Removed
+
+## [V00g] - 2024-11-06
 
 ### Added